requirements.txt
pyyaml>=6.0
faker>=18.0.0
termcolor>=2.3.0
requests>=2.31.0
requests_toolbelt>=0.9.1
jsonpath-ng>=1.6.0
lxml>=4.9.0